Split hansard analysis into query terms and analysis
